añadir al carrito

+++ app/Http/Controllers/CarritoController.php
@@ -2,10 +2,16 @@
 
 namespace App\Http\Controllers;
 
+use Illuminate\Support\Facades\Auth;
+
+
 use App\Http\Requests\StoreCarritoRequest;
 use App\Http\Requests\UpdateCarritoRequest;
 use App\Models\Carrito;
+use App\Models\Factura;
+use App\Models\Linea;
 use App\Models\User;
+use App\Models\Zapato;
 
 class CarritoController extends Controller
 {
@@ -16,27 +22,36 @@ class CarritoController extends Controller
      */
     public function index()
     {
-        $carritos = Carrito::where('user_id',auth()->user()->id)->get();
+        $carritos = Carrito::where('user_id', auth()->user()->id)->get();
         //$Totalarticulos = Carrito::where('user_id',auth()->user()->id)->sum('cantidad');
-        
-        $sumalinea=0;
+
+        $sumalinea = 0;
 
         foreach ($carritos as $key => $linea) {
             # code...
-            $sumalinea+=$linea->zapato->precio * $linea->cantidad;
+            $sumalinea += $linea->zapato->precio * $linea->cantidad;
         }
-//dd($sumalinea);
+        //dd($sumalinea);
 
-        return view('carritos/index',['carritos'=>$carritos,
-                                'totalCarrito'=> $sumalinea]);
+        return view('carritos/index', [
+            'carritos' => $carritos,
+            'totalCarrito' => $sumalinea
+        ]);
     }
 
 
     public function VaciarCarrito()
     {
-        $carritos = Carrito::where('user_id',auth()->user()->id)->get();
+        $carritos = Carrito::where('user_id', auth()->user()->id)->get();
 
-           // dd($carritos);
+        
+        if (!$carritos->count()) {
+            
+            return redirect()->route('carrito.index')->with('error', 'El carrito esta vacio.');
+       
+            // dd($carritos->count());
+        }
+        // dd($carritos);
         $carritos->each->delete();
 
         return redirect()->route('dashboard')->with('success', 'El carrito esta vacio.');
@@ -44,9 +59,30 @@ class CarritoController extends Controller
 
     public function Comprar()
     {
+        $carrito = Carrito::where('user_id', auth()->user()->id)->get();
 
 
-        return 'comprar';
+        if ($carrito->count()) {
+
+            $factura = new Factura();
+            $factura->user_id = Auth::user()->id;
+            $factura->save();
+
+
+            foreach ($carrito as $lineacarrito) {
+                $linea_nueva = new Linea();
+                $linea_nueva->factura_id = $factura->id;
+                $linea_nueva->zapato_id = $lineacarrito->zapato_id;
+                $linea_nueva->cantidad = $lineacarrito->cantidad;
+                $linea_nueva->save();
+            }
+
+            $carrito->each->delete();
+
+            return redirect()->route('zapateria')->with('success', 'Pedido realizado.');
+        } else {
+            return redirect()->route('carrito.index')->with('error', 'el carrito esta vacio.');
+        }
     }
 
 
@@ -55,6 +91,35 @@ class CarritoController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
+
+    public function agregarCarrito(Zapato $zapato)
+    {
+        //$carrito = Carrito::where('zapato_id', $zapato->id)->first();
+        
+        $carrito = Carrito::where('zapato_id', $zapato->id)->where('user_id', auth()->user()->id)->first();
+
+        //dd(Auth::user()->id.'-'.$carrito.$zapato);
+
+        //dd($zapato->id .''.$carrito->count());
+        if (!$carrito) {
+            
+            $carrito = New Carrito();
+            $carrito->user_id = Auth::user()->id;
+            $carrito->zapato_id = $zapato->id;
+            $carrito->cantidad = 1;
+        }else{
+            $carrito->cantidad =$carrito->cantidad+1;
+        }
+
+        $carrito->save();
+
+        return redirect()->route('dashboard')->with('success', 'Producto añadido con exito.');
+
+
+        
+    }
+
+
     public function create()
     {
         //
@@ -115,8 +180,4 @@ class CarritoController extends Controller
     {
         //
     }
-
-    
-
-    
 }
